refactor component structure: split logic into hooks/services, add TypeScript types
